+++ core/spec/models/spree/promotion_code/batch_builder_spec.rb
@@ -12,7 +12,7 @@ describe Spree::PromotionCode::BatchBuilder do
     )
   end
 
-  subject { described_class.new promotion_code_batch }
+  subject { described_class.new(promotion_code_batch) }
 
   describe "#build_promotion_codes" do
     context "with a failed build" do
@@ -27,6 +27,7 @@ describe Spree::PromotionCode::BatchBuilder do
         expect(promotion_code_batch.reload.state).to eq("failed")
       end
     end
+
     context "with a successful build" do
       before do
         allow(Spree::PromotionCodeBatchMailer)
@@ -48,13 +49,33 @@ describe Spree::PromotionCode::BatchBuilder do
         expect(subject.promotion.codes.map(&:value).uniq.size).to eq(10)
       end
 
+      it "updates the promotion code batch state to completed" do
+        expect(promotion_code_batch.state).to eq("completed")
+      end
+    end
+  end
+
+  describe "#join_character" do
+    context "with the default join charachter _" do
       it "builds codes with the same base prefix" do
+        subject.build_promotion_codes
+
         values = subject.promotion.codes.map(&:value)
         expect(values.all? { |val| val.starts_with?("#{base_code}_") }).to be true
       end
+    end
 
-      it "updates the promotion code batch state to completed" do
-        expect(promotion_code_batch.state).to eq("completed")
+    context "with a custom join separator" do
+      it "builds codes with the same base prefix" do
+        Spree::PromotionCode::BatchBuilder.join_character = "x"
+        subject.build_promotion_codes
+
+        values = subject.promotion.codes.map(&:value)
+        expect(values.all? { |val| val.starts_with?("#{base_code}x") }).to be true
+
+        # Reset the class join_character as it's not changed back to _ after
+        # the spec
+        Spree::PromotionCode::BatchBuilder.join_character = "_"
       end
     end
   end