+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Transform the relation between TaxRate and TaxCategory to a Many to Many [\#1851](https://github.com/solidusio/solidus/pull/1851) ([vladstoick](https://github.com/vladstoick))
 
   This fixes issue [\#1836](https://github.com/solidusio/solidus/issues/1836). By allowing a TaxRate to tax multiple categories, stores don't have to create multiple TaxRates with the same value if a zone doesn't have different tax rates for some tax categories.
+- Allow custom join character separator for promotion [\#1947](https://github.com/solidusio/solidus/pull/1947) ([vladstoick](https://github.com/vladstoick))
 
 
 ## Solidus 2.2.1 (2017-05-09)